Add wetzel-desy/mod_want_digest
As per the request from @mlassnig.

mod_want_digest is an alpha version of an Apache2 httpd module
implementing a data integrity check in accordance with RFC 3230. It
should be necessary to implement third party copy (TPC) between WebDAV
instances.

We used a dynamic dependency approach for `mod_want_digest`, this means
that the current `mod_want_digest` is curled from the git repository and
compiled within `httpd:latest`. As per @maany's request. (The other
option would have been to use static dependencies/files from
`mod_want_digest`, but this would require us to pin the `httpd` version,
and this would clash with an earlier requirement of @mlassnig.

Some additional steps are necessary to compile, initialise and attach
the module:

* Download dependencies for `libz.so` (we achieved this by installing
  `build-essential`.
* Soft-link the instance of `libz.so` to `libzlib.so` (which the linker
  `ld` can find).
* Compile using `apxs` with the option `-lzlib`.
* Initialise and attach using `apxs` in a separate step.

For SSL to work it was necessary to softlink the `default-ssl.conf` from
`sites-available` to `sites-enabled`.

Removed references to certificates

because these certificates will be volume mounted in.