servers: build openssl3.1 for alma9 daemons
Adapt the current dockerfile to always build openssl3.1 and
davix and store the new RPMs ready for installation. This is
to avoid having to build two different containers for alma9.

Add an ENV variable switch in the start-daemon.sh script to
install these RPMs, if desired, just before running rucio.