daemons/server: Add Grid CAs to the system trust
The origin of the /etc/grid-security directory appears to be Globus.
GFAL uses it transparently, but all other native system utilities and
libraries do not.

This commits adds the Grid CAs to the system trust at the time the
container starts, when applicable.  This will be important for the Rucio
components that:

  * Connect directly to the RSEs (Automatix, Dark Reaper, Reaper)
  * Communicate with FTS (Cleaner, Poller, Submitter, and the Rucio
    servers)

The /etc/grid-security directory does not exist by default; it has to be
populated externally and mounted in the container with
extraHostPathMounts.