FTS: Pin FTS to latest version used in Rucio 34.2.0
3.14.0-r2404100902git47e773b is the latest FTS version that was used
in Rucio 34.2.0, which was released on 2024-04-16.

This change is to address this issue:
https://github.com/rucio/rucio/issues/6774.

fts-monitoring has no specified versions in the fts3-token-test repo, so
it's left unpinned to latest. (I'm not even sure if this package
actually gets installed, as there's this log in the Docker build logs:
No package fts-monitoring available."