webui: update docs on `RUCIO_WEBUI_SERVER_CA_BUNDLE`
When SSL is enabled, the outbound requests from NodeJS will validate the host certificates of the server. The common CAs are pre-configured in NodeJS. However, in cases like Atlas, where the CERN CA is used, we need to explicitly provide the CA bundle that can verify the host ceritifates of the rucio server and the rucio auth server.

The documentation has been updated to reflect this behaviour.