Make NoDecode the default setting for AllowEncodedSlashes
This removes the current RUCIO_HTTPD_ENCODED_SLASHES_NO_DECODE and RUCIO_HTTPD_ENCODED_SLASHES settings so that the Apache option AllowEncodedSlashes is always set to NoDecode.

In most cases that I'm aware of, this change should be safe. The main reason we want it is so that we can use a simpler decoding method on the server to split a scope and name within a URL and get rid of some complex code. Without this option being enabled, we can't do that, because some experiments use names that start with a forward slash, and this gets merged with the separator slash by the web server.

The main implication of enabling this setting is that any encoded slashes (i.e. %2F) in URLs will not be merged with adjacent slashes. I'm not aware of any cases where Rucio relies on this happening and this has been a supported option for some time, just not the default.